% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{add_age}
\alias{add_age}
\title{Add age to a given reference date}
\usage{
add_age(birthDate, years)
}
\arguments{
\item{years}{}

\item{birthdate}{vector of dates}
}
\value{
vector of dates
}
\description{
Adds an interval (years) to a birthDate
}
