% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDM.R
\name{appointments_billings_cdm}
\alias{appointments_billings_cdm}
\title{CDM item status of patients in appointment view}
\usage{
appointments_billings_cdm(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, cdm_chosen = cdm_item_names, lazy = FALSE,
  screentag = FALSE, screentag_print = TRUE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{lazy}{if TRUE, then do not recalculate appointment list. otherwise, re-calculate}

\item{date_from=dMeasure_obj$date_a}{start date}

\item{date_to=dMeasure_obj$date_b}{end date (inclusive)}

\item{clinicians=dMeasure_obj$clinicians}{list of clinicians to view}

\item{cdm_chosen=cdm_item_names}{item types to show, defaults to all available}

\item{screentag=FALSE}{optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print=TRUE}{optionally add a 'printable' description of 'action'}
}
\value{
list of appointments (with patient details)
}
\description{
filter to CDM item billed prior to (or on) the day of displayed appointments
only show most recent billed item in each category
}
