% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{ATSI_35_44_list}
\alias{ATSI_35_44_list}
\title{list of patients who are 35 to 44 years, and ATSI, at time of $Date}
\usage{
ATSI_35_44_list(dMeasure_obj, appointments = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments}{dataframe of appointments $InternalID and $Date
if no parameter provided, derives from $appointments_filtered}
}
\value{
a vector of numbers, which are the InternalIDs
}
\description{
list of patients who are 35 to 44 years, and ATSI, at time of $Date
}
