% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{bmi30_list}
\alias{bmi30_list}
\title{list of patients with BMI>=30 (obesity)}
\usage{
bmi30_list(dMeasure_obj, appointments = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments}{list of appointments. default is $appointments_filtered

 needs appointments, as looks for recording prior to the appointment time}
}
\value{
a vector of numbers, which are the InternalIDs of patients who have
 BMI 30 or more (obesity)
}
\description{
list of patients with BMI>=30 (obesity)
}
