% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{BPdatabase}
\alias{BPdatabase}
\alias{BPdatabaseNames}
\title{show database configurations}
\usage{
BPdatabase(dMeasure_obj)

BPdatabaseNames(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object

reactive version : BPdatabaseR}

\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
dataframe of database descriptions
 id, Name, Address, Database, UserID, dbPassword

 Address will look something like "COMPUTERNAME\\BPSINSTANCE"
  note that '\' needed to be quoted, so becomes '\\'
 Database should be 'BPSPATIENTS' (or perhaps 'BPSSAMPLES')
 userID should always be 'bpsrawdata'

vector of names of database configurations
}
\description{
show database configurations

show database configuration names
}
\examples{
dMeasure_obj <- dMeasure$new()
dMeasure_obj$open_configuration_db()
dMeasure_obj$read_configuration_db()
dMeasure_obj$BPdatabase
dMeasure_obj <- dMeasure$new()
dMeasure_obj$open_configuration_db()
dMeasure_obj$read_configuration_db()
dMeasure_obj$BPdatabaseNames
}
