% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{calc_age}
\alias{calc_age}
\title{Calculate age a given reference date}
\usage{
calc_age(birthDate, refDate = Sys.Date())
}
\arguments{
\item{refDate}{reference date(s)}

\item{birthdate}{vector of dates}
}
\value{
period(s) (in years)
}
\description{
Create an interval between the date of birth and the enrollment date;
}
