% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{check_subscription}
\alias{check_subscription}
\title{check the subscription database}
\usage{
check_subscription(
  dMeasure_obj,
  clinicians = NA,
  date_from = NA,
  date_to = NA,
  adjustedate = TRUE
)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}

\item{clinicians}{vector of users to check}

\item{date_from}{date from, by default $date_a}

\item{date_to}{date to, by default $date_b}

\item{adjustdate}{will this function change the dates? ($date_a, $date_b)}
}
\value{
a list $changedate, $date_from, $date_to
 $changedate (TRUE/FALSE), and the (possibly) adjusted dates
 warning generated if $changedate is TRUE
}
\description{
check the subscription database
}
\examples{
dMeasure_obj$check_subscription()
}
