% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{choose_clinicians}
\alias{choose_clinicians}
\title{chosen clinicians}
\usage{
choose_clinicians(dMeasure_obj, choices = "", view_name = "All")
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{choices=""}{list of clinicians chosen}

\item{view_name="All"}{view}
}
\value{
list of clinicians chosen
this will be 'checked' against actual available clinicians ($clinicians_list)
}
\description{
clinicians chosen for appointment viewing
modifies self$clinicians
}
