% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{choose_location}
\alias{choose_location}
\title{Choose location}
\usage{
choose_location(dMeasure_obj, location = dMeasure_obj$location)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{location}{any of the practice locations/groups. can also be 'All'}
}
\value{
the location (after being updated, if possible)

returns an error, and does not update the location, if trying to
set to an unavailable location
}
\description{
Location is used in subsequent list of clinicians available
}
