% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{chronicliverdisease_list}
\alias{chronicliverdisease_list}
\title{list of patients with chronic liver disease}
\usage{
chronicliverdisease_list(dMeasure_obj, appointments = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments}{dataframe of appointments $InternalID and $Date

 if no parameter provided, derives from $appointments_filtered}
}
\value{
a vector of numbers, which are the InternalIDs
}
\description{
liver disease (BP doesn't have 'chronic liver disease'!), cirrhosis, alcoholism
}
