% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{chroniclungdisease_list}
\alias{chroniclungdisease_list}
\title{list of patients with chronic lung disease}
\usage{
chroniclungdisease_list(dMeasure_obj, intID)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{intID}{list of InternalID}
}
\value{
a list of numbers, which are the InternalIDs
}
\description{
list of patients with chronic lung disease
}
