% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{clinician_list}
\alias{clinician_list}
\title{find available list of clinician appointments to view}
\usage{
clinician_list(dMeasure_obj, view_name = "All")
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{view_name}{name of view. default is All i.e. 'no specific view'}
}
\value{
the clinician choice list
}
\description{
adjusts self$clinician_choice_list
according to 'view_name' (and applicable view restrictions
for the identified authenticated user)
and self$location
}
