% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{configuration_file_path}
\alias{configuration_file_path}
\title{read configuration filepaths}
\value{
SQL filepath
}
\description{
By default, the YAML configuration is either in the working
directory (where a local installation of R lives),
or the user's home directory
}
\details{
this method will set $yaml_config_filepath and $sql_config_filepath
it will read the YAML configuration filepath, which if already
existing might contain the 'real' location of the $sql_config_filepath

returns the SQL filepath
}
\examples{
dMeasure_obj <- dMeasure$new()
dMeasure_obj$configuration_file_path # read filepath
dMeasure_obj$configuration_file_path <- "c:/config.sqlite"
 # sets filepath
}
