% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{configuration_file_yaml}
\alias{configuration_file_yaml}
\title{read (or set) configuration filepath in YAML}
\usage{
configuration_file_yaml(dMeasure_obj, value)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{value}{(opt) filepath to set}
}
\value{
local_config
 $config_file : SQL filepath (only returned if no 'value' provided)
}
\description{
By default, the YAML configuration is either in the working
directory (where a local installation of R lives),
or the user's home directory
}
\details{
'~/.DailyMeasure_cfg.yaml'

this method will read or write the .sqlite filepath
to the YAML configuration file. If already
existing might contain the 'real' location of the $sql_config_filepath

Does not change the configuration file used by
the current dM object.

returns the SQL filepath
}
