% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{cvd_list}
\alias{cvd_list}
\title{list of patients with cardiovascular disease}
\usage{
cvd_list(dMeasure_obj, appointments = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments}{dataframe of appointments $InternalID and $Date
if no parameter provided, derives from $appointments_filtered}
}
\value{
a vector of numbers, which are the InternalIDs
}
\description{
for CVD risk assessment purposes
 these patients are already at high risk and
 so excluded from CVD risk assessment
}
\details{
https://www.cvdcheck.org.au/australian-absolute-cardiovascular-disease-risk-calculator
}
