% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateContact.R
\docType{data}
\name{dateContact}
\alias{dateContact}
\title{dateContact class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dateContact
}
\description{
date information for dMeasure
}
\details{
dateContact class
}
\section{Fields}{

\describe{
\item{\code{date_a}}{starting date of search (inclusive)}

\item{\code{date_b}}{ending date of search (inclusive)}
}}

\section{Methods}{

\itemize{
\item{\code{\link{choose_date}} : read (or choose) dates}
}
}

\examples{

}
\keyword{datasets}
