% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{dateformat_choice}
\alias{dateformat_choice}
\title{choose (or read) dateformat choice}
\usage{
dateformat_choice(dMeasure_obj, choice)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{choice}{name of database choice

 possible values must be one of `self$dateformat_choices`}
}
\value{
the current choice, if choice not provided
}
\description{
This must be one of `self$dateformat_choices`
}
\details{
Sets `self$dateformat` reactive, if shiny/reactive
environment available
}
\examples{
dMeasure_obj$dateformat_choice # returns the current choice
dMeasure_obj$dateformat_choice <- "2021-01-17" # sets dataformat choice
}
