% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{dateformat_function}
\alias{dateformat_function}
\title{dateformat function}
\usage{
dateformat_function(dMeasure_obj)
}
\arguments{
\item{none}{}
}
\value{
a function which formats the date to a character string
}
\description{
Returns function which formats date using `self$dateformat_choice`
}
\details{
if lubridate  package is not available, return function which
converts date to character string using default `as.character` method.

reactive version is `dateformat_functionR`, reacts to `self$dateformat()`
}
