% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{diabetes_type1_list}
\alias{diabetes_type1_list}
\title{list of patients with diabetes type 1}
\usage{
diabetes_type1_list(dMeasure_obj, appointments = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments}{dataframe of appointments $InternalID and $Date

 If no dataframe provided, derives from $appointments_filtered}
}
\value{
a vector of numbers, which are the InternalIDs
}
\description{
list of patients with diabetes type 1
}
