% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{dMeasure}
\alias{dMeasure}
\title{dMeasure class}
\description{
case-finding in EMR (Best Practice)
}
\details{
dMeasure class
}
\section{Methods}{

\itemize{
\item{\code{\link{configuration_file_path}} : read (or initiate) YAML/SQL DB filepaths}
\item{\code{\link{open_configuration_db}} : open SQLite configuration database}
\item{\code{\link{read_configuration_db}} : read SQLite configuration database}
\item{\code{\link{open_emr_db}} : open Best Practice database}
\item{\code{\link{initialize_emr_tables}} : configure Best Practice datatables}
\item{\code{\link{location_list}} : list practice locations/groups}
\item{\code{\link{choose_location}} : change, or read, current location}
}
}

\examples{

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{yaml_config_filepath}}{- filepath of YAML configuration}

\item{\code{sql_config_filepath}}{- filepath of SQL configuration (NULL is not connected)}

\item{\code{local_config}}{- in-memory copy of YAML configuration}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{local_config}}{- in-memory copy of YAML configuration}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-list_fobt}{\code{dMeasure$list_fobt()}}
\item \href{#method-list_cst}{\code{dMeasure$list_cst()}}
\item \href{#method-list_mammogram}{\code{dMeasure$list_mammogram()}}
\item \href{#method-list_zostavax}{\code{dMeasure$list_zostavax()}}
\item \href{#method-list_measlesVax}{\code{dMeasure$list_measlesVax()}}
\item \href{#method-list_influenza}{\code{dMeasure$list_influenza()}}
\item \href{#method-list_vax}{\code{dMeasure$list_vax()}}
\item \href{#method-filter_investigations}{\code{dMeasure$filter_investigations()}}
\item \href{#method-filter_investigations_appointment}{\code{dMeasure$filter_investigations_appointment()}}
\item \href{#method-filter_investigations_named}{\code{dMeasure$filter_investigations_named()}}
\item \href{#method-filter_correspondence}{\code{dMeasure$filter_correspondence()}}
\item \href{#method-filter_correspondence_appointment}{\code{dMeasure$filter_correspondence_appointment()}}
\item \href{#method-filter_correspondence_named}{\code{dMeasure$filter_correspondence_named()}}
\item \href{#method-view_incoming}{\code{dMeasure$view_incoming()}}
\item \href{#method-filter_appointments}{\code{dMeasure$filter_appointments()}}
\item \href{#method-filter_appointments_time}{\code{dMeasure$filter_appointments_time()}}
\item \href{#method-list_appointments}{\code{dMeasure$list_appointments()}}
\item \href{#method-list_visits}{\code{dMeasure$list_visits()}}
\item \href{#method-diabetes_list}{\code{dMeasure$diabetes_list()}}
\item \href{#method-asthma_list}{\code{dMeasure$asthma_list()}}
\item \href{#method-atsi_list}{\code{dMeasure$atsi_list()}}
\item \href{#method-malignancy_list}{\code{dMeasure$malignancy_list()}}
\item \href{#method-hiv_list}{\code{dMeasure$hiv_list()}}
\item \href{#method-haemoglobinopathy_list}{\code{dMeasure$haemoglobinopathy_list()}}
\item \href{#method-asplenic_list}{\code{dMeasure$asplenic_list()}}
\item \href{#method-transplant_list}{\code{dMeasure$transplant_list()}}
\item \href{#method-cvd_list}{\code{dMeasure$cvd_list()}}
\item \href{#method-cardiacdisease_list}{\code{dMeasure$cardiacdisease_list()}}
\item \href{#method-trisomy21_list}{\code{dMeasure$trisomy21_list()}}
\item \href{#method-bmi30_list}{\code{dMeasure$bmi30_list()}}
\item \href{#method-chroniclungdisease_list}{\code{dMeasure$chroniclungdisease_list()}}
\item \href{#method-neurologic_list}{\code{dMeasure$neurologic_list()}}
\item \href{#method-chronicliverdisease_list}{\code{dMeasure$chronicliverdisease_list()}}
\item \href{#method-chronicrenaldisease_list}{\code{dMeasure$chronicrenaldisease_list()}}
\item \href{#method-pregnant_list}{\code{dMeasure$pregnant_list()}}
\item \href{#method-postnatal_list}{\code{dMeasure$postnatal_list()}}
\item \href{#method-fifteenplus_list}{\code{dMeasure$fifteenplus_list()}}
\item \href{#method-sixtyfiveplus_list}{\code{dMeasure$sixtyfiveplus_list()}}
\item \href{#method-seventyfiveplus_list}{\code{dMeasure$seventyfiveplus_list()}}
\item \href{#method-ATSI_35_44_list}{\code{dMeasure$ATSI_35_44_list()}}
\item \href{#method-fortyfiveseventyfour_list}{\code{dMeasure$fortyfiveseventyfour_list()}}
\item \href{#method-cst_eligible_list}{\code{dMeasure$cst_eligible_list()}}
\item \href{#method-mammogram_eligible_list}{\code{dMeasure$mammogram_eligible_list()}}
\item \href{#method-familialHypercholesterolaemia_list}{\code{dMeasure$familialHypercholesterolaemia_list()}}
\item \href{#method-LVH_list}{\code{dMeasure$LVH_list()}}
\item \href{#method-location.insert}{\code{dMeasure$location.insert()}}
\item \href{#method-location.update}{\code{dMeasure$location.update()}}
\item \href{#method-location.delete}{\code{dMeasure$location.delete()}}
\item \href{#method-location.list}{\code{dMeasure$location.list()}}
\item \href{#method-location.permission}{\code{dMeasure$location.permission()}}
\item \href{#method-user_login}{\code{dMeasure$user_login()}}
\item \href{#method-empty_password}{\code{dMeasure$empty_password()}}
\item \href{#method-user_logout}{\code{dMeasure$user_logout()}}
\item \href{#method-password.set}{\code{dMeasure$password.set()}}
\item \href{#method-password.reset}{\code{dMeasure$password.reset()}}
\item \href{#method-server.insert}{\code{dMeasure$server.insert()}}
\item \href{#method-server.update}{\code{dMeasure$server.update()}}
\item \href{#method-server.delete}{\code{dMeasure$server.delete()}}
\item \href{#method-server.list}{\code{dMeasure$server.list()}}
\item \href{#method-server.permission}{\code{dMeasure$server.permission()}}
\item \href{#method-WriteLog}{\code{dMeasure$WriteLog()}}
\item \href{#method-userrestriction.change}{\code{dMeasure$userrestriction.change()}}
\item \href{#method-userconfig.insert}{\code{dMeasure$userconfig.insert()}}
\item \href{#method-userconfig.update}{\code{dMeasure$userconfig.update()}}
\item \href{#method-userconfig.delete}{\code{dMeasure$userconfig.delete()}}
\item \href{#method-userconfig.list}{\code{dMeasure$userconfig.list()}}
\item \href{#method-userrestriction.list}{\code{dMeasure$userrestriction.list()}}
\item \href{#method-useradmin.permission}{\code{dMeasure$useradmin.permission()}}
\item \href{#method-list_contact_appointments}{\code{dMeasure$list_contact_appointments()}}
\item \href{#method-list_contact_visits}{\code{dMeasure$list_contact_visits()}}
\item \href{#method-list_contact_services}{\code{dMeasure$list_contact_services()}}
\item \href{#method-list_contact_count}{\code{dMeasure$list_contact_count()}}
\item \href{#method-list_contact_diabetes}{\code{dMeasure$list_contact_diabetes()}}
\item \href{#method-list_contact_chroniclungdisease}{\code{dMeasure$list_contact_chroniclungdisease()}}
\item \href{#method-list_contact_15plus}{\code{dMeasure$list_contact_15plus()}}
\item \href{#method-list_contact_65plus}{\code{dMeasure$list_contact_65plus()}}
\item \href{#method-list_contact_45_74}{\code{dMeasure$list_contact_45_74()}}
\item \href{#method-list_contact_75plus}{\code{dMeasure$list_contact_75plus()}}
\item \href{#method-list_contact_ATSI_35_44}{\code{dMeasure$list_contact_ATSI_35_44()}}
\item \href{#method-list_contact_cst}{\code{dMeasure$list_contact_cst()}}
\item \href{#method-new}{\code{dMeasure$new()}}
\item \href{#method-close}{\code{dMeasure$close()}}
\item \href{#method-finalize}{\code{dMeasure$finalize()}}
\item \href{#method-open_configuration_db}{\code{dMeasure$open_configuration_db()}}
\item \href{#method-read_configuration_db}{\code{dMeasure$read_configuration_db()}}
\item \href{#method-BPdatabaseChoice_new}{\code{dMeasure$BPdatabaseChoice_new()}}
\item \href{#method-match_user}{\code{dMeasure$match_user()}}
\item \href{#method-clinician_list}{\code{dMeasure$clinician_list()}}
\item \href{#method-choose_clinicians}{\code{dMeasure$choose_clinicians()}}
\item \href{#method-open_emr_db}{\code{dMeasure$open_emr_db()}}
\item \href{#method-initialize_emr_tables}{\code{dMeasure$initialize_emr_tables()}}
\item \href{#method-choose_location}{\code{dMeasure$choose_location()}}
\item \href{#method-choose_date}{\code{dMeasure$choose_date()}}
\item \href{#method-list_allergy}{\code{dMeasure$list_allergy()}}
\item \href{#method-list_socialHx}{\code{dMeasure$list_socialHx()}}
\item \href{#method-list_familyHx}{\code{dMeasure$list_familyHx()}}
\item \href{#method-list_dataQuality}{\code{dMeasure$list_dataQuality()}}
\item \href{#method-influenzaVax_obs}{\code{dMeasure$influenzaVax_obs()}}
\item \href{#method-HbA1C_obs}{\code{dMeasure$HbA1C_obs()}}
\item \href{#method-smoking_obs}{\code{dMeasure$smoking_obs()}}
\item \href{#method-BloodPressure_obs}{\code{dMeasure$BloodPressure_obs()}}
\item \href{#method-UrineAlbumin_obs}{\code{dMeasure$UrineAlbumin_obs()}}
\item \href{#method-PersistentProteinuria_obs}{\code{dMeasure$PersistentProteinuria_obs()}}
\item \href{#method-eGFR_obs}{\code{dMeasure$eGFR_obs()}}
\item \href{#method-Cholesterol_obs}{\code{dMeasure$Cholesterol_obs()}}
\item \href{#method-clone}{\code{dMeasure$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_fobt"></a>}}
\subsection{Method \code{list_fobt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_fobt(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  action = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_cst"></a>}}
\subsection{Method \code{list_cst()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_cst(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  action = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_mammogram"></a>}}
\subsection{Method \code{list_mammogram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_mammogram(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  action = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_zostavax"></a>}}
\subsection{Method \code{list_zostavax()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_zostavax(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  vaxtag = FALSE,
  vaxtag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_measlesVax"></a>}}
\subsection{Method \code{list_measlesVax()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_measlesVax(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  vaxtag = FALSE,
  vaxtag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_influenza"></a>}}
\subsection{Method \code{list_influenza()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_influenza(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  vaxtag = FALSE,
  vaxtag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_vax"></a>}}
\subsection{Method \code{list_vax()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_vax(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  vaxtag = FALSE,
  vaxtag_print = TRUE,
  chosen = self$vaccine_choices
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_investigations"></a>}}
\subsection{Method \code{filter_investigations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$filter_investigations(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  Action = NA,
  Actioned = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_investigations_appointment"></a>}}
\subsection{Method \code{filter_investigations_appointment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$filter_investigations_appointment(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  Action = NA,
  Actioned = NA,
  ignorePast = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_investigations_named"></a>}}
\subsection{Method \code{filter_investigations_named()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$filter_investigations_named(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  Action = NA,
  Actioned = NA,
  ignorePast = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_correspondence"></a>}}
\subsection{Method \code{filter_correspondence()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$filter_correspondence(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  Action = NA,
  Actioned = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_correspondence_appointment"></a>}}
\subsection{Method \code{filter_correspondence_appointment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$filter_correspondence_appointment(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  Action = NA,
  Actioned = NA,
  ignorePast = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_correspondence_named"></a>}}
\subsection{Method \code{filter_correspondence_named()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$filter_correspondence_named(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  Action = NA,
  Actioned = NA,
  ignorePast = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-view_incoming"></a>}}
\subsection{Method \code{view_incoming()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$view_incoming(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  Action = NA,
  Actioned = NA,
  ignorePast = NA,
  lazy = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_appointments"></a>}}
\subsection{Method \code{filter_appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$filter_appointments(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  status = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_appointments_time"></a>}}
\subsection{Method \code{filter_appointments_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$filter_appointments_time(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  status = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_appointments"></a>}}
\subsection{Method \code{list_appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_appointments(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  status = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_visits"></a>}}
\subsection{Method \code{list_visits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_visits(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  visit_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-diabetes_list"></a>}}
\subsection{Method \code{diabetes_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$diabetes_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asthma_list"></a>}}
\subsection{Method \code{asthma_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$asthma_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-atsi_list"></a>}}
\subsection{Method \code{atsi_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$atsi_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-malignancy_list"></a>}}
\subsection{Method \code{malignancy_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$malignancy_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hiv_list"></a>}}
\subsection{Method \code{hiv_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$hiv_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-haemoglobinopathy_list"></a>}}
\subsection{Method \code{haemoglobinopathy_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$haemoglobinopathy_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asplenic_list"></a>}}
\subsection{Method \code{asplenic_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$asplenic_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transplant_list"></a>}}
\subsection{Method \code{transplant_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$transplant_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cvd_list"></a>}}
\subsection{Method \code{cvd_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$cvd_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cardiacdisease_list"></a>}}
\subsection{Method \code{cardiacdisease_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$cardiacdisease_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trisomy21_list"></a>}}
\subsection{Method \code{trisomy21_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$trisomy21_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bmi30_list"></a>}}
\subsection{Method \code{bmi30_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$bmi30_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chroniclungdisease_list"></a>}}
\subsection{Method \code{chroniclungdisease_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$chroniclungdisease_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-neurologic_list"></a>}}
\subsection{Method \code{neurologic_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$neurologic_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chronicliverdisease_list"></a>}}
\subsection{Method \code{chronicliverdisease_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$chronicliverdisease_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chronicrenaldisease_list"></a>}}
\subsection{Method \code{chronicrenaldisease_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$chronicrenaldisease_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pregnant_list"></a>}}
\subsection{Method \code{pregnant_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$pregnant_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-postnatal_list"></a>}}
\subsection{Method \code{postnatal_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$postnatal_list(
  appointments = NULL,
  include_edc = FALSE,
  days_min = 0,
  days_max = 180,
  outcome = c(0:7)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fifteenplus_list"></a>}}
\subsection{Method \code{fifteenplus_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$fifteenplus_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sixtyfiveplus_list"></a>}}
\subsection{Method \code{sixtyfiveplus_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$sixtyfiveplus_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seventyfiveplus_list"></a>}}
\subsection{Method \code{seventyfiveplus_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$seventyfiveplus_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ATSI_35_44_list"></a>}}
\subsection{Method \code{ATSI_35_44_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$ATSI_35_44_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fortyfiveseventyfour_list"></a>}}
\subsection{Method \code{fortyfiveseventyfour_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$fortyfiveseventyfour_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cst_eligible_list"></a>}}
\subsection{Method \code{cst_eligible_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$cst_eligible_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mammogram_eligible_list"></a>}}
\subsection{Method \code{mammogram_eligible_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$mammogram_eligible_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-familialHypercholesterolaemia_list"></a>}}
\subsection{Method \code{familialHypercholesterolaemia_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$familialHypercholesterolaemia_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LVH_list"></a>}}
\subsection{Method \code{LVH_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$LVH_list(appointments = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-location.insert"></a>}}
\subsection{Method \code{location.insert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$location.insert(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-location.update"></a>}}
\subsection{Method \code{location.update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$location.update(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-location.delete"></a>}}
\subsection{Method \code{location.delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$location.delete(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-location.list"></a>}}
\subsection{Method \code{location.list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$location.list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-location.permission"></a>}}
\subsection{Method \code{location.permission()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$location.permission()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_login"></a>}}
\subsection{Method \code{user_login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$user_login(password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empty_password"></a>}}
\subsection{Method \code{empty_password()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$empty_password()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_logout"></a>}}
\subsection{Method \code{user_logout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$user_logout()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-password.set"></a>}}
\subsection{Method \code{password.set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$password.set(newpassword, oldpassword = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-password.reset"></a>}}
\subsection{Method \code{password.reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$password.reset(user, newpassword = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server.insert"></a>}}
\subsection{Method \code{server.insert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$server.insert(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server.update"></a>}}
\subsection{Method \code{server.update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$server.update(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server.delete"></a>}}
\subsection{Method \code{server.delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$server.delete(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server.list"></a>}}
\subsection{Method \code{server.list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$server.list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server.permission"></a>}}
\subsection{Method \code{server.permission()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$server.permission()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WriteLog"></a>}}
\subsection{Method \code{WriteLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$WriteLog(message)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-userrestriction.change"></a>}}
\subsection{Method \code{userrestriction.change()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$userrestriction.change(restriction, state)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-userconfig.insert"></a>}}
\subsection{Method \code{userconfig.insert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$userconfig.insert(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-userconfig.update"></a>}}
\subsection{Method \code{userconfig.update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$userconfig.update(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-userconfig.delete"></a>}}
\subsection{Method \code{userconfig.delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$userconfig.delete(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-userconfig.list"></a>}}
\subsection{Method \code{userconfig.list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$userconfig.list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-userrestriction.list"></a>}}
\subsection{Method \code{userrestriction.list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$userrestriction.list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-useradmin.permission"></a>}}
\subsection{Method \code{useradmin.permission()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$useradmin.permission()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_appointments"></a>}}
\subsection{Method \code{list_contact_appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_appointments(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  status = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_visits"></a>}}
\subsection{Method \code{list_contact_visits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_visits(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  visit_type = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_services"></a>}}
\subsection{Method \code{list_contact_services()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_services(date_from = NA, date_to = NA, clinicians = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_count"></a>}}
\subsection{Method \code{list_contact_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_count(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  contact_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_diabetes"></a>}}
\subsection{Method \code{list_contact_diabetes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_diabetes(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  contact_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_chroniclungdisease"></a>}}
\subsection{Method \code{list_contact_chroniclungdisease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_chroniclungdisease(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  contact_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_15plus"></a>}}
\subsection{Method \code{list_contact_15plus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_15plus(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  contact_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_65plus"></a>}}
\subsection{Method \code{list_contact_65plus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_65plus(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  contact_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_45_74"></a>}}
\subsection{Method \code{list_contact_45_74()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_45_74(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  contact_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_75plus"></a>}}
\subsection{Method \code{list_contact_75plus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_75plus(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  contact_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_ATSI_35_44"></a>}}
\subsection{Method \code{list_contact_ATSI_35_44()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_ATSI_35_44(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  contact_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_contact_cst"></a>}}
\subsection{Method \code{list_contact_cst()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_contact_cst(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  contact_type = NA,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open_configuration_db"></a>}}
\subsection{Method \code{open_configuration_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$open_configuration_db(
  configuration_file_path = self$configuration_file_path
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_configuration_db"></a>}}
\subsection{Method \code{read_configuration_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$read_configuration_db(config_db = self$config_db)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BPdatabaseChoice_new"></a>}}
\subsection{Method \code{BPdatabaseChoice_new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$BPdatabaseChoice_new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-match_user"></a>}}
\subsection{Method \code{match_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$match_user()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clinician_list"></a>}}
\subsection{Method \code{clinician_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$clinician_list(view_name = "All", location = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-choose_clinicians"></a>}}
\subsection{Method \code{choose_clinicians()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$choose_clinicians(choices = "", view_name = "All")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open_emr_db"></a>}}
\subsection{Method \code{open_emr_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$open_emr_db(BPdatabaseChoice = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initialize_emr_tables"></a>}}
\subsection{Method \code{initialize_emr_tables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$initialize_emr_tables(emr_db = self$emr_db)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-choose_location"></a>}}
\subsection{Method \code{choose_location()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$choose_location(location = self$location)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-choose_date"></a>}}
\subsection{Method \code{choose_date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$choose_date(date_from = self$date_a, date_to = self$date_b)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_allergy"></a>}}
\subsection{Method \code{list_allergy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_allergy(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  qualitytag = FALSE,
  qualitytag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_socialHx"></a>}}
\subsection{Method \code{list_socialHx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_socialHx(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  qualitytag = FALSE,
  qualitytag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_familyHx"></a>}}
\subsection{Method \code{list_familyHx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_familyHx(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  qualitytag = FALSE,
  qualitytag_print = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_dataQuality"></a>}}
\subsection{Method \code{list_dataQuality()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$list_dataQuality(
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  qualitytag = FALSE,
  qualitytag_print = TRUE,
  chosen = self$dataQuality_choices
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-influenzaVax_obs"></a>}}
\subsection{Method \code{influenzaVax_obs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$influenzaVax_obs(intID, date_from = NA, date_to = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HbA1C_obs"></a>}}
\subsection{Method \code{HbA1C_obs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$HbA1C_obs(intID, date_from = NA, date_to = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-smoking_obs"></a>}}
\subsection{Method \code{smoking_obs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$smoking_obs(intID, date_from = NA, date_to = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BloodPressure_obs"></a>}}
\subsection{Method \code{BloodPressure_obs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$BloodPressure_obs(intID, date_from = NA, date_to = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UrineAlbumin_obs"></a>}}
\subsection{Method \code{UrineAlbumin_obs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$UrineAlbumin_obs(intID, date_from = NA, date_to = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersistentProteinuria_obs"></a>}}
\subsection{Method \code{PersistentProteinuria_obs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$PersistentProteinuria_obs(intID, date_from = NA, date_to = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eGFR_obs"></a>}}
\subsection{Method \code{eGFR_obs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$eGFR_obs(intID, date_from = NA, date_to = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cholesterol_obs"></a>}}
\subsection{Method \code{Cholesterol_obs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$Cholesterol_obs(intID, date_from = NA, date_to = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasure$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
