% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{.identified_user}
\alias{.identified_user}
\title{returns information about the identified user}
\usage{
.identified_user(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}
}
\value{
dataframe, single row, the identified user
}
\description{
also sets reactive $identified_user, and sets $authenticated
}
