% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{eGFR_obs}
\alias{eGFR_obs}
\title{List of eGFR observations/recordings}
\usage{
eGFR_obs(dMeasure_obj, intID, date_from = NA, date_to = NA)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{intID}{vector of InternalID}

\item{date_from}{start date. default is $date_b minus 12 months}

\item{date_to}{end date (inclusive). default is $date_b}
}
\value{
dataframe of InternalID, BPDate, BP
 BP will be a string of two numbers separated by "/"
}
\description{
Filtered by InternalID (vector patient numbers) and dates
}
\details{
the reference date for 'most recent' measurement is 'date_to'
}
