% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_password.R
\name{empty_password}
\alias{empty_password}
\title{is password for the current identified user available?}
\usage{
empty_password(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}
}
\value{
TRUE if password is not available
}
\description{
returns TRUE if password is not set (NA) or empty
}
