% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appointments.R
\name{filter_appointments}
\alias{filter_appointments}
\title{List of appointments}
\usage{
filter_appointments(dMeasure_obj, date_from = NULL, date_to = NULL,
  clinicians = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from=dMeasure_obj$date_a}{start date}

\item{date_to=dMeasure_obj$date_b}{end date (inclusive)}

\item{clinicians=dMeasure_obj$clinicians}{list of clinicians to view}
}
\value{
list of appointments
}
\description{
Filtered by date, and chosen clinicians
Stores result in $appointments_filtered
}
