% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appointments.R
\name{filter_appointments_time}
\alias{filter_appointments_time}
\title{List of appointments with 'human (and R)' readable time-of-day}
\usage{
filter_appointments_time(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, status = NA, lazy = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{(default $date_a) start date, inclusive (date object)}

\item{date_to}{(default $date_b) end date, inclusive (date object)}

\item{clinicians}{(default $clinicians list) of clinicians to view}

\item{status}{(default NA) filter by 'status' if not NA
permissible values are 'Booked', 'Completed', 'At billing',
'Waiting', 'With doctor'}

\item{lazy}{(default FALSE) if lazy=TRUE, then don't re-calculate $appointments_filtered to calculate}
}
\value{
list of appointments
}
\description{
Filtered by date, and chosen clinicians
Stores result in $appointments_filtered_time
}
