% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative_incoming.R
\name{filter_correspondence}
\alias{filter_correspondence}
\title{List of correspondence}
\usage{
filter_correspondence(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, Action = NA, Actioned = NA)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from=dMeasure_obj$date_a}{start date}

\item{date_to=dMeasure_obj$date_b}{end date (inclusive)}

\item{clinicians=dMeasure_obj$clinicians}{list of clinicians to view}

\item{Action=NA}{Filter by action?
a vector of actions (in string form) or NULL
e.g. "Urgent Appointment" and/or "Non-urgent Appointment" or "No action"
if NA, will adopt the value of $filter_incoming_Action}

\item{Actioned=NA}{Filter by having been 'actioned?' i.e. notified
can be logical (TRUE or FALSE), a NULL
or a Date (actioned prior to or by 'Actioned' Date)
if NA, will adopt the value of $filter_incoming_Actioned}
}
\value{
list of correspondence
}
\description{
Filtered by date, and chosen clinicians
Stores result in $correspondence_filtered
}
