% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative.R
\name{filter_investigations}
\alias{filter_investigations}
\title{List of investigations}
\usage{
filter_investigations(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, Action = NULL, Actioned = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from=dMeasure_obj$date_a}{start date}

\item{date_to=dMeasure_obj$date_b}{end date (inclusive)}

\item{clinicians=dMeasure_obj$clinicians}{list of clinicians to view}

\item{Action=NULL}{Filter by action?
a vector of actions (in string form)
e.g. "Urgent Appointment" and/or "Non-urgent Appointment" or "No action"}

\item{Actioned=NULL}{Filter by having been 'actioned?' i.e. notified
can be logical (TRUE or FALSE)
or a Date (actioned prior to or by 'Actioned' Date)}
}
\value{
list of investigations
}
\description{
Filtered by date, and chosen clinicians
Stores result in $investigations_filtered
}
