% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerScreen_bowelcancer.R
\name{fobt_list}
\alias{fobt_list}
\title{Bowel cancer screening list}
\usage{
fobt_list(dMeasure_obj, appointments_list = NA, emr_db = NA,
  action = TRUE, screentag = FALSE, screentag_print = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments_list}{dataframe, list of appointments to search
needs Age (presumably at time of appointment),
      InternalID (the EMR's identification code fo the patient)}

\item{emr_db}{R6 object, accesss to Best Practice (EMR) database}

\item{action=TRUE}{includes 'OutOfDate' field}

\item{screentag=FALSE}{optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print=FALSE}{optionally add a 'printable' description of 'action'}
}
\value{
list of appointments (with patient details)
 adds the following fields
  TestDate (date object) - date and
  TestName - description of the most recent bowel cancer screening test (if any)
  OutOfDateTest - 1 = never done, 2 = overdue, 3 = 'up-to-date'

  unlike the original version of this function, no HTML
}
\description{
Bowel cancer screening list
}
