% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{formatdate}
\alias{formatdate}
\title{formatdate function}
\usage{
formatdate(dMeasure_obj)
}
\arguments{
\item{none}{}
}
\value{
a function which formats the date to a character string
}
\description{
Returns function which formats date using `self$dateformat_choice`
}
\details{
if lubridate  package is not available, return function which
converts date to character string using default `as.character` method.

reactive version is `formatdateR`, reacts to `self$dateformat()`
}
