% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{framingham_riskequation}
\alias{framingham_riskequation}
\title{Framingham Risk Equation}
\usage{
framingham_riskequation(df, years = 5, outcome = "CVD")
}
\arguments{
\item{df}{a dataframe, which must include the following columns/fields...

 InternalID (numeric)

 CardiovascularDisease (logical)

 Diabetes (logical)

 SmokingDate (date) - not actually used in this equation

 SmokingStatus (chr) - "Smoker" if a smoker. All other values ignored

 UrineAlbuminDate (chr), UrineAlbuminValue (double), UrineAlbuminUnit (chr)

 PersistentProteinuria (logical)

 eGFRDate (date), eGFRValue (double), eGFRUnits (chr)

 FamilialHypercholesterolaemia (logical)

 LVH (logical) = left ventricular hyp0ertrophy

 CholesterolDate (date), Cholesterol (double), HDL (double), LDL (double),
  Triglycerides (double), CholHDLRatio (double)

 BPDate (date), BP (character, two numbers separated by "/")

 Sex (character), Ethnicity (character)

 Age (double)}

\item{years}{number of years to predict (from 4 to 12). default is 5 years.}

\item{outcome}{(default is "CVD") "CHD" - coronary heart disease. includes myocardial
 infarction, death from coronary heart disease plus angina pectoris and
 coronary insufficiency.

 "CVD" cardiovascular disease (the default) includes coronary heart disease,
 stroke (including transient ischaemia), congestive heart failure and
 peripheral vascular disease.}
}
\value{
dataframe

 InternalID

 frisk : numeric (a number, or 'NA' if not enough information to computer)

 friskHI : either 'NA' or '>15\%'

 '>15\%' are groups considered equivalent risk to already having IHD
}
\description{
Calculate cardiovascular disease risk according to Framingham Risk Equation
}
\details{
sourced from National Vascular Disease Prevention Alliance (Australia)

http://cvdcheck.org.au/pdf/Absolute_CVD_Risk-Quick_Reference_Guide.pdf

"An Updated Coronary Risk Profile - A Statement for Health Professionals"
Keaven M. Anderson, Peter W.F. Wilson, Patricia M. Odell, William B. Kannel

AHA (American Heart Association) Medical/Scientific Statement

sourced from http://ahajournals.org

and "Cardiovascular Disease Risk Profiles"
Keaven M. Anderson, Patricia M. Odell, Peter W.F. Wilson, William B. Kannel
American Heart Journal : 1991; 121:293-298
}
\examples{

framingham_riskequation(data.frame(InternalID = 1, BP = "135/80", Sex = "Female",
 Age = 55,
 SmokingStatus = "Smoker", CholHDLRatio = 230/48, Diabetes = TRUE, LVH = FALSE,
  CardiovascularDisease = FALSE, PersistentProteinuria = FALSE, eGFRValue = NA,
   eGFRUnits = NA, UrineAlbuminValue = NA, UrineAlbuminUnits = NA,
   FamilialHypercholesterolaemia = NA, Cholesterol = 5.96, Ethnicity = NA),
    outcome = "CHD", years = 10)
    # this comes from "Cardiovascular disease risk profiles" (Anderson 1991)
    # the worked answer in the paper is 0.22. this function returns 0.2189125
    # the same risk-factors with outcome = "CVD" and years = 5 returns 0.180
    #  (cvdcheck.org.au reports 18\%)
    # the same risk-factors except Sex = "Male" with outcome = "CVD" and years = 5 returns 0.202
    #  (cvdcheck.org.au reports 20\%)

framingham_riskequation(data.frame(InternalID = 2, BP = "130/80", Sex = "Male",
 Age = 55,
 SmokingStatus = "Smoker", CholHDLRatio = 240/45, Diabetes = FALSE, LVH = FALSE,
  CardiovascularDisease = FALSE, PersistentProteinuria = FALSE, eGFRValue = NA,
   eGFRUnits = NA, UrineAlbuminValue = NA, UrineAlbuminUnits = NA,
   FamilialHypercholesterolaemia = NA, Cholesterol = 6.22, Ethnicity = NA),
    outcome = "CHD", years = 10)
    # this comes from "An Updated Coronary Risk Profile" (Anderson 1991)
    # the worked answer in the paper is 0.192, this function returns 0.1919
    # the same risk-factors with outcome = "CVD" and years = 5 returns 0.133
    #  (cvdcheck.org.au reports 13\%)
    # the same risk-factors except LVH = TRUE (outcome = CVD, =years 5) returns 0.211
    #  (cvdcheck.org.au reports 21\%)

framingham_riskequation(data.frame(InternalID = 3, BP = "130/80", Sex = "Female",
 Age = 55,
 SmokingStatus = "Smoker", CholHDLRatio = 240/45, Diabetes = FALSE, LVH = FALSE,
  CardiovascularDisease = FALSE, PersistentProteinuria = FALSE, eGFRValue = NA,
   eGFRUnits = NA, UrineAlbuminValue = NA, UrineAlbuminUnits = NA,
   FamilialHypercholesterolaemia = NA, Cholesterol = 6.22, Ethnicity = NA),
    outcome = "CHD", years = 10)
    # this comes from "An Updated Coronary Risk Profile" (Anderson 1991)
    # the worked answer in the paper is 0.135, this function returns 0.1349
    # the same risk-factors with outcome = "CVD" and years = 5 returns 0.088
    #  (cvdcheck.org.au reports 9\%)
    # the same risk-factors except LVH = TRUE (outcome = CVD, =years 5) returns 0.150
    #  (cvdcheck.org.au reports 15\%)

}
