% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{HbA1C_obs}
\alias{HbA1C_obs}
\title{List of HbA1C observations/recordings}
\usage{
HbA1C_obs(dMeasure_obj, IntID, date_from = NA, date_to = NA)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{IntID}{vector of InternalID}

\item{date_from}{start date. default is $date_b minus 12 months}

\item{date_to}{end date (inclusive). default is $date_b}
}
\value{
dataframe of InternalID, HbA1CDate, HbA1CValue, HbA1CUnits
}
\description{
Filtered by InternalID (vector patient numbers) and dates
}
\details{
the reference date for 'most recent' measurement is 'date_to'
}
