% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{hrmin}
\alias{hrmin}
\title{Calculate seconds to a 'time' starting from midnight}
\usage{
hrmin(t)
}
\arguments{
\item{t}{seconds}
}
\value{
24-hour time of form '14:15' (hh:mm)
}
\description{
t: value in seconds
}
