% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative_incoming.R
\name{incoming_view}
\alias{incoming_view}
\title{incoming (investigations and correspondence) view}
\usage{
incoming_view(
  dMeasure_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  Action = NA,
  Actioned = NA,
  ignorePast = NA,
  lazy = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{(default dMeasure_obj$date_a) start date}

\item{date_to}{(default dMeasure_obj$date_b) end date (inclusive)}

\item{clinicians}{(default dMeasure_obj$clinicians) list of clinicians to view}

\item{Action}{(default NA) Filter by action?
a vector of actions (in string form) or NULL
e.g. "Urgent Appointment" and/or "Non-urgent Appointment" or "No action"
if NA, will adopt the value of $filter_incoming_Action}

\item{Actioned}{(default NA) Filter by having been 'actioned?' i.e. notified
can be logical (TRUE or FALSE), a NULL
or a Date (actioned prior to or by 'Actioned' Date)
if NA, will adopt the value of $filter_incoming_Actioned}

\item{ignorePast}{(default NA) ignore appointments before current Sys.Date() date?
retrieves from $filter_incoming_ignorePast if NA}

\item{lazy}{if TRUE, then do not recalculate appointment list. otherwise, re-calculate}

\item{screentag}{(default FALSE) optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print}{(default TRUE) optionally add a 'printable' description of 'action'}
}
\value{
list of appointments (with patient details)
}
\description{
filter to date range
and clinicians
}
