% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{initialize_data_table}
\alias{initialize_data_table}
\title{initialize_data_table}
\usage{
initialize_data_table(config_db, tablename, variable_list)
}
\arguments{
\item{config_db}{R6 object of configuration database}

\item{tablename}{name of table}

\item{variable_list}{list of variable headings, with variable type
e.g. list(c("id", "integer"), c("Name", "character"))}
}
\value{
nothing
}
\description{
write tables to configuration database
}
\details{
Make sure the table in the database has all the right variable headings.
  Creates table if table does not yet exist.
  Allows 'update' of old databases.
  Alters table in configuration database directly.

  Used by `$open_configuration_db`
}
