% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{initialize_emr_tables}
\alias{initialize_emr_tables}
\title{initialize the tables of the EMR database}
\usage{
initialize_emr_tables(dMeasure_object, emr_db = dMeasure_object$emr_db)
}
\arguments{
\item{dMeasure_object}{dMeasure object}

\item{emr_db}{R6 object connecting to EMR database}
}
\description{
initialize the tables of the EMR database
}
