% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{interval}
\alias{interval}
\title{Calculate period between two dates}
\usage{
interval(date_a, date_b, unit = "none")
}
\arguments{
\item{date_a}{vector of dates}

\item{date_b}{vector of dates}

\item{unit}{"none" or "month". if "month", convert all years tomonths}
}
\value{
period(s) in $year, $month and $day
}
\description{
Create an interval between the date_a and date_b
}
\details{
can return 'negative' numbers
returns NA if either of date_a or date_b is NA
returns an arbitrarily large number of years (Inf) if min(date_a, date_b) is -Inf
}
