% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appointments.R
\name{list_appointments}
\alias{list_appointments}
\title{List of appointments with date of birth}
\usage{
list_appointments(dMeasure_obj, date_from = NULL, date_to = NULL,
  clinicians = NULL, lazy = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from=dMeasure_obj$date_a}{start date, inclusive (date object)}

\item{date_to=dMeasure_obj$date_b}{end date, inclusive (date object)}

\item{clinicians=dMeasure_obj$clinicians}{list of clinicians to view}

\item{lazy=FALSE}{if lazy=TRUE, then don't re-calculate $appointments_filtered to calculate}
}
\value{
list of appointments
}
\description{
Filtered by date, and chosen clinicians
}
