% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{list_asthma_details}
\alias{list_asthma_details}
\title{list_asthma}
\usage{
list_asthma_details(
  dMeasure_obj,
  contact = FALSE,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  ignoreOld = FALSE,
  include_uptodate = TRUE,
  lazy = FALSE
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{contact}{contact system (if TRUE) or appointment system (if FALSE)}

\item{date_from}{by default, dM$date_a}

\item{date_to}{by default, dM$date_b}

\item{clinicians}{by default, is dM$clinicians}

\item{min_contact}{used for the 'contact' system. minimum number of contacts in time period}

\item{min_date}{used for the 'contact' system. must have been seen since min_date}

\item{max_date}{used for the 'contact' system. must have been before the max_date}

\item{contact_type}{used for the 'contact' system}

\item{ignoreOld}{ignore old administrations (>15 months from date_to).
by default, FALSE}

\item{include_uptodate}{remove uptodate vaccinations?}

\item{lazy}{lazy evaluation?}
}
\value{
a dataframe
 structure of dataframe depends on whether 'contact' or 'appointment' system
 in both cases, the columns includes :
   Patient, InternalID, DOB, RecordNo, FluvaxName, FluvaxDate
 in the case of 'appointment' system, also includes appointment details
}
\description{
list_asthma
}
