% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{list_contact_appointments}
\alias{list_contact_appointments}
\title{List of contacts (appointment type)}
\usage{
list_contact_appointments(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, status = NA)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{(default NA -> date_a field) start date}

\item{date_to}{(default NA -> date_b field) end date (inclusive)}

\item{clinicians}{(default NA -> clinicians field) list of clinicians to view}

\item{status}{(default NA) filter by 'status' if not NA
permissible values are 'Booked', 'Completed', 'At billing',
'Waiting', 'With doctor'
if NA, adopts from active $appointment_status}
}
\value{
dataframe of Patient (name), InternalID, AppointmentDate
}
\description{
Filtered by date, and chosen clinicians
}
