% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{list_contact_count}
\alias{list_contact_count}
\title{List of contacts counts}
\usage{
list_contact_count(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, min_contact = 1, min_date = as.Date(-Inf, origin =
  "1970-01-01"), lazy = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{start date. default is $dateContact$date_a}

\item{date_to}{end date (inclusive). default is $dateContact$date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{min_contact}{minimum number of contacts. default is one (1)}

\item{min_date}{most recent contact must be at least min_date. default is -Inf}
}
\value{
dataframe of Patient (name), InternalID, Counts
}
\description{
Filtered by date, and chosen clinicians
}
