% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactConditions.R
\name{list_contact_diabetes}
\alias{list_contact_diabetes}
\title{List of diabetics in the contact list}
\usage{
list_contact_diabetes(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, min_contact = NA, min_date = NA, lazy = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{start date. default is $dateContact$date_a}

\item{date_to}{end date (inclusive). default is $dateContact$date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{min_contact}{minimum number of contacts. default is $contact_min, initially one (1)}

\item{min_date}{most recent contact must be at least min_date. default is $contact_minDate, initially -Inf}
}
\value{
dataframe of Patient (name), InternalID, Count, and most recent contact date
}
\description{
Filtered by date, and chosen clinicians
}
