% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{list_contact_services}
\alias{list_contact_services}
\title{List of contacts (service type)}
\usage{
list_contact_services(
  dMeasure_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{(default NA -> dMeasure_obj$date_a) start date}

\item{date_to}{(default NA -> dMeasure_obj$date_b) end date (inclusive)}

\item{clinicians}{(default NA -> dMeasure_obj$clinicians) list of clinicians to view}
}
\value{
dataframe of Patient (name), InternalID, ServiceDate
}
\description{
Filtered by date, and chosen clinicians
}
