% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{list_contact_visits}
\alias{list_contact_visits}
\title{List of contacts (visit type)}
\usage{
list_contact_visits(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, visit_type = NA)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from=dMeasure_obj$date_a}{start date}

\item{date_to=dMeasure_obj$date_b}{end date (inclusive)}

\item{clinicians=dMeasure_obj$clinicians}{list of clinicians to view}

\item{visit_types=NA}{filter by 'visit_type' if not NA
permissible values are "Home", "Non Visit", "Hospital", "RACF",
"Telephone", "SMS", "Email", "Locum Service", "Out of Office",
"Other", "Hostel", "Telehealth"}
}
\value{
dataframe of Patient (name), InternalID, VisitDate
}
\description{
Filtered by date, and chosen clinicians
}
