% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerScreen.R
\name{list_cst}
\alias{list_cst}
\title{Cervical screening list with tags}
\usage{
list_cst(
  dMeasure_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  include_uptodate = TRUE,
  lazy = FALSE,
  action = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{from date range (default \verb{$date_a})}

\item{date_to}{to date range (default \verb{$date_b})}

\item{clinicians}{list of clinicians (default \verb{$clinicians})}

\item{appointments_list}{dataframe, list of appointments to search
\itemize{
\item if not provided, use \verb{$appointments_list}
\item needs fields \code{Age}, \code{InternalID}, \code{AppointmentDate}
}}

\item{include_uptodate}{include those who are up-to-date ('green' tags)}

\item{lazy}{recalculate an appointment list}

\item{action}{includes 'OutOfDate' field}

\item{screentag}{optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print}{optionally add a 'printable' description of 'action'}
}
\value{
list of appointments (with patient details)
\itemize{
\item adds \code{OutofDateTest} (if \code{action} is \code{TRUE}) and \code{screentag}/\code{screentag_print}
}
}
\description{
Cervical screening list with tags
}
