% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerScreen.R
\name{list_cst_details}
\alias{list_cst_details}
\title{Cervical screening list details}
\usage{
list_cst_details(
  dMeasure,
  screen_cst_list,
  screen_cst_id = NULL,
  include_uptodate = TRUE
)
}
\arguments{
\item{screen_cst_list}{appointment list, minimum columns}

\item{screen_cst_id}{vector of applicable IDs, if already found
if not provided, will be calculated from \code{screen_cst_list}}

\item{include_uptodate}{include those who are up-to-date ('green' tags)}
}
\value{
returns the screen_cst_list dataframe with added columns, including...
\describe{
\item{TestDate}{(date object) - date}
\item{TestName}{description of the most recent cervical cancer screening test (if any)}
\item{OutOfDateTest}{1 = never done, 2 = overdue, 3 = 'up-to-date'}
}
and many other result detail columns.
}
\description{
Cervical screening list details
}
