% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQuality.R
\name{list_dataQuality}
\alias{list_dataQuality}
\title{List of patients for data quality check}
\usage{
list_dataQuality(
  dMeasure_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  lazy = FALSE,
  qualitytag = FALSE,
  qualitytag_print = TRUE,
  chosen = vax_names
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{from date range (default $date_a)}

\item{date_to}{to date range (default $date_b)}

\item{clinicians}{list of clinicians (default $clinicians)}

\item{appointments_list}{provide an appointment list (default $appointments_list)}

\item{lazy}{(default FALSE) recalculate an appointment list}

\item{qualitytag}{(default FALSE) HTML/browser version of tags}

\item{qualitytag_print}{(default TRUE) printable version of tags}

\item{chosen}{list of data quality names (default is all)}
}
\value{
dataframe list of influenza eligible patients
}
\description{
Optionally added a HTML ('qualitytag') or printable ('qualitytag_print')
}
