% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQuality.R
\name{list_familyHx}
\alias{list_familyHx}
\title{List of patients with family history recordings}
\usage{
list_familyHx(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, appointments_list = NULL, lazy = FALSE,
  qualitytag = FALSE, qualitytag_print = TRUE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{from date range (default $date_a)}

\item{date_to}{to date range (default $date_b)}

\item{clinicians}{list of clinicians (default $clinicians)}

\item{appointments_list}{provide an appointment list (default $appointments_list)}

\item{lazy}{= FALSE recalculate an appointment list}

\item{qualitytag}{= FALSE}

\item{qualitytag_print}{= TRUE}
}
\value{
dataframe list of patients with family history recording status
}
\description{
Optionally added a HTML ('qualitytag') or printable ('qualitytag_print')
}
