% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerScreen.R
\name{list_fobt}
\alias{list_fobt}
\title{Bowel cancer screening list}
\usage{
list_fobt(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, appointments_list = NULL, lazy = FALSE,
  action = FALSE, screentag = FALSE, screentag_print = TRUE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{from date range (default 'date_a' field)}

\item{date_to}{to date range (default 'date_b' field)}

\item{clinicians}{list of clinicians (default 'clinicians' field)}

\item{appointments_list}{dataframe, list of appointments to search
(as opposed to using 'appointments_list' method)
needs Age (presumably at time of appointment),
      InternalID (the EMR's identification code fo the patient)}

\item{lazy}{= FALSE recalculate an appointment list}

\item{action}{= FALSE includes 'OutOfDate' field}

\item{screentag}{= FALSE optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print}{= TRUE optionally add a 'printable' description of 'action'}
}
\value{
list of appointments (with patient details)
 adds the following fields
  TestDate (date object) - date and
  TestName - description of the most recent bowel cancer screening test (if any)
  OutOfDateTest - 1 = never done, 2 = overdue, 3 = 'up-to-date'
}
\description{
Bowel cancer screening list
}
