% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunization.R
\name{list_influenza}
\alias{list_influenza}
\title{List of patients potentially eligible for influenza vaccine}
\usage{
list_influenza(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, appointments_list = NULL, lazy = FALSE,
  vaxtag = FALSE, vaxtag_print = TRUE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{from date range (default $date_a)}

\item{date_to}{to date range (default $date_b)}

\item{clinicians}{list of clinicians (default $clinicians)}

\item{appointments_list}{provide an appointment list (default $appointments_list)}

\item{lazy}{= FALSE recalculate an appointment list}

\item{vaxtag}{= FALSE}

\item{vaxtag_print}{= TRUE}
}
\value{
dataframe list of influenza eligible patients
}
\description{
Includes patients who may have already had influenza vaccine,
date of previous influenza vaccine is included.
}
\details{
Optionally added a HTML ('vaxtag') or printable ('vaxtag_print')
}
