% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerScreen.R
\name{list_mammogram}
\alias{list_mammogram}
\title{Breast cancer  screening list}
\usage{
list_mammogram(
  dMeasure_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  appointments_list = NULL,
  include_uptodate = TRUE,
  lazy = FALSE,
  action = FALSE,
  screentag = FALSE,
  screentag_print = TRUE
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{from date range (default $date_a)}

\item{date_to}{to date range (default $date_b)}

\item{clinicians}{list of clinicians (default $clinicians)}

\item{appointments_list}{dataframe, list of appointments to search

 if not provided, use $appointments_list

 needs fields Age, InternalID}

\item{lazy}{recalculate an appointment list}

\item{action}{includes 'OutOfDate' field}

\item{screentag}{optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print}{optionally add a 'printable' description of 'action'}
}
\value{
list of appointments (with patient details)
 adds the following fields

 \describe{
  \item{TestDate}{(date object) - date}
  \item{TestName}{description of the most recent breast cancer screening test (if any)}
  \item{OutOfDateTest}{1 = never done, 2 = overdue, 3 = 'up-to-date'}
 }
}
\description{
Breast cancer  screening list
}
