% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityImprovementMeasures.R
\name{list_qim_diabetes}
\alias{list_qim_diabetes}
\title{List of diabetics, with Quality Improvement Measures, in the contact list}
\usage{
list_qim_diabetes(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, min_contact = NA, min_date = NA, ignoreOld = NA,
  lazy = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{start date. default is $dateContact$date_a}

\item{date_to}{end date (inclusive). default is $dateContact$date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{min_contact}{minimum number of contacts. default is $contact_min, initially one (1)}

\item{min_date}{most recent contact must be at least min_date. default is $contact_minDate, initially -Inf}

\item{ignoreOld}{ignore results/observatioins that don't qualify for quality improvement measures
if not supplied, reads $qim_ignoreOld}

\item{lazy}{recalculate the diabetes contact list?}
}
\value{
dataframe of Patient (name), InternalID, Count, and most recent contact date
}
\description{
Filtered by date, and chosen clinicians
}
\details{
QIM 01 - HbA1C - most recent. the QIM measure is within last twelve months
QIM 05 - Influenza immunization - most recent. the QIM measure is within last 15 months
QIM 10 - Blood pressure - most recent. the QIM measure is within the last six monthe

the reference date for 'most recent' measurement is 'date_to'
}
