% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billings.R
\name{list_services}
\alias{list_services}
\title{List of services with date of birth}
\usage{
list_services(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, screentag = FALSE, screentag_print = TRUE,
  lazy = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{default is $date_a. start date, inclusive (date object)}

\item{date_to}{default is $date_b end date, inclusive (date object)}

\item{clinicians}{default is $clinicians. list of clinicians to view}

\item{screentag}{(default FALSE) optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print}{(default TRUE) optionally add a 'printable' description of 'action'}

\item{lazy}{if lazy=TRUE, then don't re-calculate $appointments_filtered to calculate}
}
\value{
list of services
}
\description{
Filtered by date, and chosen clinicians
}
