% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appointments.R
\name{list_visits}
\alias{list_visits}
\title{List of visits with date of birth}
\usage{
list_visits(
  dMeasure_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  visit_type = NA,
  lazy = FALSE
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{default is $date_a. start date, inclusive (date object)}

\item{date_to}{default is $date_b end date, inclusive (date object)}

\item{clinicians}{default is $clinicians. list of clinicians to view}

\item{visit_type}{default is $visit_type. filter by 'visittype' if not NA}

\item{lazy}{if lazy=TRUE, then don't re-calculate $appointments_filtered to calculate}
}
\value{
list of visits
}
\description{
Filtered by date, and chosen clinicians
}
