% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunization.R
\name{list_zostavax}
\alias{list_zostavax}
\title{List of patients potentially eligible for Zostavax}
\usage{
list_zostavax(
  dMeasure_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  intID = NULL,
  intID_Date = Sys.Date(),
  appointments_list = NULL,
  include_uptodate = TRUE,
  lazy = FALSE,
  vaxtag = FALSE,
  vaxtag_print = TRUE
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{from date range (default $date_a)}

\item{date_to}{to date range (default $date_b)}

\item{clinicians}{list of clinicians (default $clinicians)}

\item{intID}{list of internal ID (default is NULL, in which case appointments_list is used)}

\item{intID_Date}{if intID is not NULL, then date to check (default is Sys.Date())}

\item{appointments_list}{provide an appointment list (default $appointments_list)}

\item{include_uptodate}{include those who are up-to-date ('green' tags)}

\item{lazy}{= FALSE recalculate an appointment list}

\item{vaxtag}{= FALSE}

\item{vaxtag_print}{= TRUE}
}
\value{
dataframe list of Zostavax eligible patients
}
\description{
Includes patients who may have already had Zostavax,
date of previous Zostavax is included.
}
\details{
Optionally added a HTML ('vaxtag') or printable ('vaxtag_print')
}
