% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_location.R
\name{location.delete}
\alias{location.delete}
\title{location.delete}
\usage{
location.delete(dMeasure_obj, description)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{description}{list object : $Name}
}
\value{
dataframe - full list of location descriptions
 can also return error (stop) if deletion if cannot delete
}
\description{
delete a location description
}
