% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_location.R
\name{location.insert}
\alias{location.insert}
\title{location.insert}
\usage{
location.insert(dMeasure_obj, description)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{description}{list object : $Name, $Description}
}
\value{
dataframe - full list of location descriptions
 can also return error (stop) if description is invalid
}
\description{
insert a new location description
}
\examples{
\dontrun{
a <- dMeasure::dMeasure$new()
}
\dontrun{
a$open_emr_db()
}
\dontrun{
a$location.inser(list(Name = "Bayside", Description = "by the sea"))
}
}
