% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_location.R
\name{location.list}
\alias{location.list}
\title{location.list}
\usage{
location.list(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
dataframe - full list of location descriptions
}
\description{
list location description
}
