% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_location.R
\name{location.update}
\alias{location.update}
\title{location.update}
\usage{
location.update(dMeasure_obj, description)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{description}{list object : $id, $Name, $Description}
}
\value{
dataframe - full list of location descriptions
 can also return error (stop) if description is invalid
}
\description{
update a location description
}
