% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{location_list}
\alias{location_list}
\title{Show list of locations}
\usage{
location_list(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
the list of locations, including 'All'
}
\description{
This includes 'All'
}
