% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{mammogram_eligible_list}
\alias{mammogram_eligible_list}
\title{list of patients who are breast cancer screening eligible at time of $Date}
\usage{
mammogram_eligible_list(dMeasure_obj, appointments = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments}{dataframe of appointments $InternalID and $Date

 if no parameter provided, derives from $appointments_filtered}
}
\value{
a vector of numbers, which are the InternalIDs
}
\description{
\itemize{
 \item age fifty to seventy-four years inclusive
 \item female
}
}
\details{
https://www.cancer.org.au/about-cancer/early-detection/screening-programs/breast-cancer-screening.html
}
