% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{match_user}
\alias{match_user}
\title{Match user with current 'identified' system user}
\usage{
match_user(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}
}
\value{
self
}
\description{
Matches 'dMeasure_obj$.UserConfig$AuthIdentity' with Sys.info()[["user"]]
}
