% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{parent_list}
\alias{parent_list}
\title{list of patients who are listed as the head of family
with a child between defined ages}
\usage{
parent_list(dMeasure_obj, appointments = NULL, months_min = 0, months_max = 12)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments}{dataframe of appointments $InternalID and $Date}

\item{months_min}{minimum age in months (inclusive, 'from')}

\item{months_max}{maximum age in months (exclusive, 'up to')

 if no parameter provided, derives from $appointments_filtered}
}
\value{
a vector of numbers, which are the InternalIDs
}
\description{
This is found through the use of 'HeadOfFamilyID' in
db$patientsRaw
}
\details{
Note that this does *not* include postnatal_list, and
so a separate check is required if requiring postnatal
}
