% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_password.R
\name{password.set}
\alias{password.set}
\title{Set password of currently identified user}
\usage{
password.set(dMeasure_obj, newpassword, oldpassword = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}

\item{newpassword}{the new password}

\item{oldpassword=NULL}{the old password (if one exists)}
}
\value{
TRUE if password is successfully set
otherwise, stops with error
}
\description{
if there is an old password, that must be specified
if there is no old password, then 'oldpassword' does not need to be defined
}
