% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{paste2}
\alias{paste2}
\title{paste which can ignore NA and empty strings}
\usage{
paste2(..., sep = " ", collapse = NULL, na.rm = FALSE)
}
\arguments{
\item{...}{the list of strings to paste}

\item{sep}{the separator string, " " by default}

\item{collapse}{the collapse string, NULL by default}

\item{na.rm}{whether to remove NA and empty strings}
}
\value{
string
}
\description{
Acts the same as regular paste, unless na.rm = TRUE,
in which case empty strings and NA are removed
}
\details{
based on code by Moody_Mudskipper at
https://stackoverflow.com/questions/13673894/suppress-nas-in-paste
with additional code from
https://stackoverflow.com/questions/14270950/suppress-separator-in-paste-when-values-are-missing
}
