% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{postnatal_list}
\alias{postnatal_list}
\title{list of patients who are pregnant at the 'appointment' date}
\usage{
postnatal_list(
  dMeasure_obj,
  appointments = NULL,
  include_edc = FALSE,
  days_min = 0,
  days_max = 180,
  outcome = c(0, 1, 2, 3, 4, 5, 6, 7)
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments}{dataframe of appointments $InternalID and $Date}

\item{include_edc}{include post-natal by EDC (TRUE or FALSE)}

\item{days_min}{minimum number of days post-natal}

\item{days_max}{maximum number of days post-natal}

\item{outcome}{allowable outcomes (vector)
  possible outcomes 0 = none recorded, 1 = "Live birth",
  2 = Miscarriage, 3 = Termination, 4 = Ectopic,
  5 = IUFD (intra-uterine fetal death), 6 = stillbirth
  7 = hydatiform mole

 if no parameter provided, derives from $appointments_filtered}
}
\value{
a dataframe : InternalID, EDCbyDate, EDCbyScan, EndDate, OutcomeCode
}
\description{
list of patients who are pregnant at the 'appointment' date
}
