% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{read_configuration_db}
\alias{read_configuration_db}
\title{read the SQL configuration database}
\usage{
read_configuration_db(dMeasure_object,
  config_db = dMeasure_object$config_db)
}
\arguments{
\item{dMeasure_object}{dMeasure object}

\item{config_db}{R6 object to open SQL database - default is the internally stored value}
}
\description{
read the SQL configuration database
}
