% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{read_configuration_filepaths}
\alias{read_configuration_filepaths}
\title{read configuration filepaths}
\usage{
read_configuration_filepaths(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}
}
\value{
nothing, modifies \code{dMeasure_obj}
}
\description{
By default, the YAML configuration is either in the working
directory (where a local installation of R lives),
or the user's home directory
}
\details{
this method will set $yaml_config_filepath and $sql_config_filepath
it will read the YAML configuration filepath, which if already
existing might contain the 'real' location of the $sql_config_filepath

finally sets $configuration_file_path to 'real' SQL filepath
}
\examples{
dMeasure_obj <- dMeasure$new()
dMeasure_obj$read_configuration_filepaths()

read_configuration_filepaths(dMeasure_obj)
}
