% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{read_subscription_db}
\alias{read_subscription_db}
\title{read the subscription database}
\usage{
read_subscription_db(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}
}
\description{
read the subscription database
}
\examples{
dMeasure_obj$read_subscription_db()
}
