% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityImprovementMeasures.R
\name{report_qim_copd}
\alias{report_qim_copd}
\title{Chronic Obstructive Pulmonary Disease Quality Improvement Measure report, in the contact list}
\usage{
report_qim_copd(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, min_contact = NA, contact_type = NA,
  min_date = NA, demographic = NA, ignoreOld = NA, lazy = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{start date. default is $date_a}

\item{date_to}{end date (inclusive). default is $date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{min_contact}{minimum number of contacts. default is $contact_min, initially one (1)}

\item{contact_type}{contact types which are accepted. default is $contact_type}

\item{min_date}{most recent contact must be at least min_date. default is $contact_minDate, initially -Inf}

\item{demographic}{demographic groupings for reporting.
if not supplied, reads $qim_demographicGroup
list of available demographic groups in $qim_demographicGroupings}

\item{ignoreOld}{ignore results/observatioins that don't qualify for quality improvement measures
if not supplied, reads $qim_ignoreOld}

\item{lazy}{recalculate the copd contact list?}
}
\value{
dataframe of Patient (name), demographic, measure (done or not), Count, Proportion
}
\description{
Filtered by date, and chosen clinicians
}
\details{
Shows chosen QIM measures, and by demographic grouping

QIM 06 -Proportion of patients with COPD who were immunised against influenza

the reference date for 'most recent' measurement is 'date_to'
}
