% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_server.R
\name{server.delete}
\alias{server.delete}
\title{servers.delete}
\usage{
server.delete(description)
}
\arguments{
\item{description}{$Name}

\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
dataframe - full list of database descriptions
 can also return error (stop) if description is invalid
}
\description{
remove a server descriptioin
}
