% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_server.R
\name{server.insert}
\alias{server.insert}
\title{server.insert}
\usage{
server.insert(description)
}
\arguments{
\item{description}{$Name, $Address, $Database, $UserID, $dbPassword}

\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
dataframe - full list of database descriptions
 can also return error (stop) if description is invalid
}
\description{
insert a new server description
}
