% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_server.R
\name{server.list}
\alias{server.list}
\title{server.list}
\usage{
server.list(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
dataframe - full list of database descriptions
 can also return error (stop) if description is invalid
}
\description{
List server description. does not include passwords!
If 'ServerAdmin' restriction is set, only 'ServerAdmin'
users can see the server list with this method.
}
