% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_server.R
\name{server.update}
\alias{server.update}
\title{server.update}
\usage{
server.update(description)
}
\arguments{
\item{description}{$id, $Name, $Address, $Database, $UserID, $dbPassword}

\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
dataframe - full list of database descriptions
 can also return error (stop) if description is invalid
}
\description{
change (update) a server descriptioin
}
