% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{seventyfiveplus_list}
\alias{seventyfiveplus_list}
\title{list of patients who are seventy-five years or more in age at time of $Date}
\usage{
seventyfiveplus_list(dMeasure_obj, appointments = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments}{dataframe of appointments $InternalID and $Date
if no parameter provided, derives from $appointments_filtered}
}
\value{
a list of numbers, which are the InternalIDs
}
\description{
list of patients who are seventy-five years or more in age at time of $Date
}
