% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{simple_decode_base64}
\alias{simple_decode_base64}
\title{Simple decoder base64}
\usage{
simple_decode_base64(msg, key = NULL, nonce = NULL)
}
\arguments{
\item{msg}{the text to decode}

\item{key}{the cipher, which can be set manually, otherwise will read from env}

\item{nonce}{a non-secret unique data value used to randomize the cipher}
}
\value{
- the encrypted text

  returns NA for elements of msg which are NA, or "" empty string.
  note that simple_encode will ENCRYPT an empty string "".

  returns NULL if failure to decode e.g. invalid base64 decode
}
\description{
Simple decoder of text strings, will output a text string.
Uses sodium library and base64_enc/dec from jsonlite. Has some defaults, but
will also take command-line arguments or read from environment.
Companion function to simple_encode
}
