% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{update_date}
\alias{update_date}
\title{Update date}
\usage{
update_date(dMeasure_obj, date_from = dMeasure_obj$date_a,
  date_to = dMeasure_obj$date_b)
}
\arguments{
\item{date_from}{'From' date. default is current date_from}

\item{date_to}{'To' date. default is current date_to}
}
\value{
list(date_a, date_b)

if date_a is later than date_b, an error is returned
}
\description{
Sets 'from' and 'to' dates used in subsequent searches
}
