% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{update_subscription}
\alias{update_subscription}
\title{Update subscription database}
\usage{
update_subscription(dMeasure_obj, Fullname = NA, License = NA,
  Identifier = NA, verify = TRUE)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}

\item{Fullname}{name of the user to change license}

\item{License}{the (undecoded) license string}

\item{Identifier}{the identifier of the user}

\item{verify}{verify before changing}
}
\value{
TRUE if license written, FALSE if not
 only meaningful if Verify is TRUE, in which case
 FALSE indicates the license was not valid
}
\description{
Update subscription database
}
