% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{update_UserFullConfig}
\alias{update_UserFullConfig}
\title{update the UserFullConfig}
\usage{
update_UserFullConfig(dMeasure_object, db = dMeasure_object$db,
  UserConfig = dMeasure_object$UserConfig)
}
\arguments{
\item{dMeasure_object}{dMeasure object}

\item{db}{tables in the EMR database}

\item{UserConfig}{UserConfig stored in configuration file}
}
\value{
UserFullConfig
also updates this object's copy
}
\description{
integrates the UserConfig in the SQLite configuration file
with the user information in the EMR database
}
