% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{user_login}
\alias{user_login}
\title{check password against the current identified user}
\usage{
user_login(dMeasure_obj, password)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}

\item{password}{the password}
}
\value{
TRUE if password is correct
otherwise stops with error
}
\description{
check password against the current identified user
}
