% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{user_logout}
\alias{user_logout}
\title{Logout current identified user}
\usage{
user_logout(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}

\item{password}{the password}
}
\value{
authentication status
 error (stop) if no identified user
 warning if not authenticated (logged in)
}
\description{
Logout current identified user
}
