% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_user.R
\name{useradmin.permission}
\alias{useradmin.permission}
\title{useradmin.permission}
\usage{
useradmin.permission(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
TRUE or FALSE
 additionally returns a warning if permission is FALSE
}
\description{
Does the current user have user admin access permission?
}
\details{
Can only be 'false' if $UserRestrictions$Restriction contains
'UserAdmin'

if 'UserAdmin' in the $UserRestrictions, then a user needs
to be identified and 'authenticated'. Authentication requires
identification (via Sys.info()$user), and might also require
a password
}
