% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_user.R
\name{userconfig.delete}
\alias{userconfig.delete}
\title{userconfig.delete}
\usage{
userconfig.delete(dMeasure_obj, description)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{description}{list $Fullname}
}
\value{
$UserConfig
}
\description{
delete user configuration
}
\details{
will not delete user configuration if that user is the only holder
of an attribute which is restricted, but essential for operation
e.g. ServerAdmin
}
