% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_user.R
\name{userconfig.insert}
\alias{userconfig.insert}
\title{userconfig.insert}
\usage{
userconfig.insert(dMeasure_obj, description)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{description}{list $Fullname, $AuthIdentity, $Location, $Attributes}
}
\value{
$UserConfig
}
\description{
add user configuration
}
