% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_user.R
\name{userconfig.list}
\alias{userconfig.list}
\title{userconfig.list}
\usage{
userconfig.list(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
$UserConfig
}
\description{
list user configuration
}
