% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{UserConfig}
\alias{UserConfig}
\title{show user configurations}
\usage{
UserConfig(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
dataframe of user configuration descriptions
 id, Fullname, AuthIdentity, Location, Attributes, License

 Fullname - Best Practice full user name
 AuthIdentity - Windows login identity
 Location - vector of groups/locations
 Attributes - vector of user's attributes/permissions
 License - undecoded license
}
\description{
show user configurations
}
\examples{
dMeasure_obj <- dMeasure$new()
dMeasure_obj$open_configuration_db()
dMeasure_obj$read_configuration_db()
dMeasure_obj$UserConfig
}
