% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_user.R
\name{userconfig.update}
\alias{userconfig.update}
\title{userconfig.update}
\usage{
userconfig.update(dMeasure_obj, description)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{description}{list $Fullname, $AuthIdentity, $Location, $Attributes
$License, $LicenseCheckDate}
}
\value{
self$UserConfig
}
\description{
update user configuration
}
\details{
if setting multiple attributes or locations, preferably set the attribute
or locations as a list e.g.

list(Fullname = "Mrs. Diabetes Educator",
  Location = list(c("Jemena", "Lakeside")))

note that the list does not need to enclose with I(),
since 'as_tibble' is used for translation. If attributes or location is
not a list, userconfig.insert will turn them into a list

The description to update is defined by $Fullname
}
\examples{

 dMeasure_obj <- dMeasure::dMeasure$new()
 dMeasure_obj$open_emr_db()
 dMeasure_obj$userconfig.update(
   list(Fullname = "Mrs. Diabetes Educator",
        Location = list(c("Jemena", "Lakeside")),
        Attributes = list(c("UserAdmin", "ServerAdmin")))
}
