% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{UserFullConfig}
\alias{UserFullConfig}
\title{UserfullConfig}
\description{
integrates the UserConfig in the SQLite configuration file
with the user information in the EMR database
}
\details{
contains user names attached to configuration information
contains ALL user names
$UserConfig() contains just names who have been configured
}
