% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_user.R
\name{userrestriction.change}
\alias{userrestriction.change}
\title{userrestriction.change}
\usage{
userrestriction.change(dMeasure_obj, restriction, state)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{restriction}{name of restriction}

\item{state}{new state (TRUE or FALSE)}
}
\value{
list (newstate, error, warn)  : returns 'state', if permissible
                                      error$message and error$title
                                      warn$message and warn$title
}
\description{
Change user restrictions
}
