% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_user.R
\name{userrestriction.list}
\alias{userrestriction.list}
\title{userrestriction.list}
\usage{
userrestriction.list(dMeasure_obj)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}
}
\value{
vector of user restrictions
}
\description{
List user restrictions in use
}
