% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunization.R
\name{vax_list}
\alias{vax_list}
\title{List of patients potentailly eligible for vaccines}
\usage{
vax_list(dMeasure_obj, date_from = NA, date_to = NA, clinicians = NA,
  appointments_list = NULL, lazy = FALSE, vaxtag = FALSE,
  vaxtag_print = TRUE, chosen = vax_names)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{from date range (default self$date_a)}

\item{date_to}{to date range (default self$date_b)}

\item{clinicians}{list of clinicians (default self$clinicians)}

\item{appointments_list}{provide an appointment list (as opposed to using self$appointments_list)}

\item{lazy}{= FALSE recalculate an appointment list}

\item{vaxtag}{= FALSE}

\item{vaxtag_print}{= TRUE}

\item{chosen}{list of vaccine names (default is all)}
}
\value{
dataframe list of influenza eligible patients
}
\description{
Includes patients who may have already had the vaccine,
date of previous vaccine is included.
}
\details{
Optionally added a HTML ('vaxtag') or printable ('vaxtag_print')
}
