% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{verify_license}
\alias{verify_license}
\title{verify license/subscription}
\usage{
verify_license(License, Identifier)
}
\arguments{
\item{License}{an encoded character string}

\item{Identifier}{a character string}
}
\value{
a date object 'LicenseDate'. returns NA if not valid
}
\description{
Location is used in subsequent list of clinicians available
}
