% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_server.R
\name{WriteLog}
\alias{WriteLog}
\title{WriteLog}
\usage{
WriteLog(dMeasure_obj, message)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{message}{}
}
\value{
nothing
}
\description{
writes to Logfile (if available, and logging is turned on)
}
