% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billings.R
\name{billings_list}
\alias{billings_list}
\title{Billings done on same day as appointments}
\usage{
billings_list(dMeasure_obj, date_from, date_to, clinicians, lazy,
  screentag, screentag_print)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{lazy}{if TRUE, then do not recalculate appointment list. otherwise, re-calculate}

\item{date_from=dMeasure_obj$date_a}{start date}

\item{date_to=dMeasure_obj$date_b}{end date (inclusive)}

\item{clinicians=dMeasure_obj$clinicians}{list of clinicians to view}

\item{screentag=FALSE}{optionally add a fomantic/semantic HTML description of 'action'}

\item{screentag_print=TRUE}{optionally add a 'printable' description of 'action'}
}
\value{
list of appointments (with patient details)
}
\description{
filter to billings which are done on the same day as displayed appointments
billings are aggregated/group to patient/day
}
