% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{BPdatabaseChoice}
\alias{BPdatabaseChoice}
\title{choose (or read) database choice}
\value{
the current database choice
}
\description{
this must be one of 'None' or one of the defined databases
}
\examples{
dMeasure_obj$BPdatabaseChoice # returns the current choice
dMeasure_obj$BPdatabaseChoice <- "None" # sets database to none
}
