% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{cardiacdisease_list}
\alias{cardiacdisease_list}
\title{list of patients with cardiac conditions}
\usage{
cardiacdisease_list(dMeasure_obj, appointments = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{appointments}{dataframe of appointments $InternalID and $Date
if no parameter provided, derives from $appointments_filtered}
}
\value{
a list of numbers, which are the InternalIDs
}
\description{
for influenza immunization purposes
}
