% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{chronicliverdisease_list}
\alias{chronicliverdisease_list}
\title{list of patients with chronic liver disease}
\usage{
chronicliverdisease_list(dMeasure_obj, intID)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{intID}{list of InternalID}
}
\value{
a list of numbers, which are the InternalIDs
}
\description{
list of patients with chronic liver disease
}
