% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\docType{data}
\name{dMeasure}
\alias{dMeasure}
\title{dMeasure class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
dMeasure
}
\description{
case-finding in EMR (Best Practice)
}
\details{
dMeasure class
}
\section{Fields}{

\describe{
\item{\code{yaml_config_filepath}}{- filepath of YAML configuration}

\item{\code{sql_config_filepath}}{- filepath of SQL configuration (NULL is not connected)}

\item{\code{local_config}}{- in-memory copy of YAML configuration}
}}

\section{Methods}{

\itemize{
\item{\code{\link{read_configuration_filepaths}} : read (or initiate) YAML/SQL DB filepaths}
\item{\code{\link{open_configuration_db}} : open SQLite configuration database}
\item{\code{\link{read_configuration_db}} : read SQLite configuration database}
\item{\code{\link{open_emr_db}} : open Best Practice database}
\item{\code{\link{initialize_emr_tables}} : configure Best Practice datatables}
\item{\code{\link{update_UserFullConfig}} : create 'full' user configuratio table}
\item{\code{\link{update_date}} : change, or read, search date range}
\item{\code{\link{location_list}} : list practice locations/groups}
\item{\code{\link{update_location}} : change, or read, current location}
}
}

\examples{

}
\keyword{datasets}
