% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appointments.R
\name{filter_appointments}
\alias{filter_appointments}
\title{List of appointments}
\usage{
filter_appointments(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, status = NA)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{start date. default is $dateContact$date_a}

\item{date_to}{end date (inclusive). default is $dateContact$date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{status}{filter by 'status' if not NA
permissible values are 'Booked', 'Completed', 'At billing',
'Waiting', 'With doctor'
default is all possible appointment status
 $appointment_status_types}
}
\value{
list of appointments
}
\description{
Filtered by date, and chosen clinicians
Stores result in $appointments_filtered
}
