% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative_incoming.R
\name{filter_investigations}
\alias{filter_investigations}
\title{List of investigations}
\usage{
filter_investigations(
  dMeasure_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  Action = NA,
  Actioned = NA
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{(default dMeasure_obj$date_a) start date}

\item{date_to}{(default dMeasure_obj$date_b) end date (inclusive)}

\item{clinicians}{(default dMeasure_obj$clinicians) list of clinicians to view}

\item{Action}{(default NA) Filter by action?
a vector of actions (in string form) or NULL
e.g. "Urgent Appointment" and/or "Non-urgent Appointment" or "No action"
if NA, will adopt the value of dMeasure_obj$filter_incoming_Action
if NULL, then no Actions are chosen!}

\item{Actioned}{(default NA) Filter by having been 'actioned?' i.e. notified
can be logical (TRUE or FALSE), a NULL
or a Date (actioned prior to or by 'Actioned' Date)
if NA, will adopt the value of dMeasure_obj$filter_incoming_Actioned}
}
\value{
list of investigations
}
\description{
Filtered by date, and chosen clinicians
Stores result in $investigations_filtered
}
