% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative_incoming.R
\name{filter_investigations_appointment}
\alias{filter_investigations_appointment}
\title{List of investigations with appointments}
\usage{
filter_investigations_appointment(dMeasure_obj, date_from = NA,
  date_to = NA, clinicians = NA, Action = NA, Actioned = NA,
  lazy = FALSE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from=dMeasure_obj$date_a}{start date}

\item{date_to=dMeasure_obj$date_b}{end date (inclusive)}

\item{clinicians=dMeasure_obj$clinicians}{list of clinicians to view}

\item{Action=NA}{Filter by action?
a vector of actions (in string form) or NULL
e.g. "Urgent Appointment" and/or "Non-urgent Appointment" or "No action"
if NA, will adopt the value of $filter_incoming_Action}

\item{Actioned=NA}{Filter by having been 'actioned?' i.e. notified
can be logical (TRUE or FALSE), a NULL
or a Date (actioned prior to or by 'Actioned' Date)
if NA, will adopt the value of $filter_incoming_Actioned}

\item{lazy=FALSE}{if TRUE, don't 'call' filter_investigations
to re-calculate}
}
\value{
list of investigations with relevant appointments
}
\description{
Filtered by date, and chosen clinicians
added information about appointments scheduled after the 'Checked' date
Stores result in $investigations_filtered_appointments
}
