% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_definitions.R
\name{framingham_riskequation}
\alias{framingham_riskequation}
\title{Framingham Risk Equation}
\usage{
framingham_riskequation(df, years = 5)
}
\arguments{
\item{df}{a dataframe

 Patient (chr = character)

 InternalID (numeric)

 CardiovascularDisease (logical)

 Diabetes (logical)

 SmokingDate (date) - not actually used in this equation

 SmokingStatus (chr) - "Smoker" if a smoker. All other values ignored

 UrineAlbuminDate (chr), UrineAlbuminValue (double), UrineAlbuminUnit (chr)

 PersistentProteinuria (logical)

 eGFRDate (date), eGFRValue (double), eGFRUnits (chr)

 FamilialHypercholesterolaemia (logical)

 LVH (logical) = left ventricular hyp0ertrophy

 CholesterolDate (date), Cholesterol (double), HDL (double), LDL (double),
  Triglycerides (double), CholHDLRatio (double)

 BPDate (date), BP (character, two numbers separated by "/")

 Sex (character), Ethnicity (character)

 RecordNo (character), MaritalStatus (character), Sexuality (character)

 DOB (character), Age (double)}

\item{years}{number of years to predict (from 4 to 12)}
}
\value{
dataframe

 InternalID

 frisk : numeric

 friskHI : either 'NA' or '>15\%'

 '>15\%' are groups considered equivalent risk to already having IHD
}
\description{
Calculate cardiovascular disease risk according to Framingham Risk Equation
}
\details{
sourced from National Vascular Disease Prevention Alliance (Australia)

http://cvdcheck.org.au/pdf/Absolute_CVD_Risk-Quick_Reference_Guide.pdf

and "An Updated Coronary Risk Profile - A Statement for Health Professionals"
Keaven M. Anderson, Peter W.F. Wilson, Patricia M. Odell, William B. Kannel

AHA (American Heart Association) Medical/Scientific Statement

sourced from http://ahajournals.org
}
