% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactConditions.R
\name{list_contact_cst}
\alias{list_contact_cst}
\title{List of those eligible (as of 'date_to') for cervical screeningin the contact list}
\usage{
list_contact_cst(
  dMeasure_obj,
  date_from = NA,
  date_to = NA,
  clinicians = NA,
  min_contact = NA,
  min_date = NA,
  max_date = NA,
  contact_type = NA,
  lazy = FALSE,
  store = TRUE
)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{start date. default is $date_a}

\item{date_to}{end date (inclusive). default is $date_b}

\item{clinicians}{list of clinicians to view. default is $clinicians}

\item{min_contact}{minimum number of contacts. default is $contact_min, initially one (1)}

\item{min_date}{most recent contact must be at most max_date. default is $contact_maxDate, initially Sys.Date()}

\item{contact_type}{contact types which are accepted. default is $contact_type}

\item{lazy}{force re-calculate?}

\item{store}{keep result in  self$contact_cst_list}
}
\value{
dataframe of Patient (name), InternalID, Count, and most recent contact date
}
\description{
Age 25 to 74 years inclusive (as of 'date_to')
Female
No history of hysterectomy (except sub-total hysterectomy)
-- currently does not take regard of the recorded date of the hysterectomy
}
\details{
Filtered by date, and chosen clinicians
}
