% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunization.R
\name{list_zostavax}
\alias{list_zostavax}
\title{List of patients potentially eligible for Zostavax}
\usage{
list_zostavax(dMeasure_obj, date_from = NA, date_to = NA,
  clinicians = NA, appointments_list = NULL, lazy = FALSE,
  vaxtag = FALSE, vaxtag_print = TRUE)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{date_from}{from date range (default self$date_a)}

\item{date_to}{to date range (default self$date_b)}

\item{clinicians}{list of clinicians (default self$clinicians)}

\item{appointments_list}{provide an appointment list (as opposed to using self$appointments_list)}

\item{lazy}{= FALSE recalculate an appointment list}

\item{vaxtag}{= FALSE}

\item{vaxtag_print}{= TRUE}
}
\value{
dataframe list of Zostavax eligible patients
}
\description{
Includes patients who may have already had Zostavax,
date of previous Zostavax is included.
}
\details{
Optionally added a HTML ('vaxtag') or printable ('vaxtag_print')
}
