% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{open_configuration_db}
\alias{open_configuration_db}
\title{Open the SQL connection to the configuration from the SQL configuration file}
\usage{
open_configuration_db(dMeasure_obj,
  configuration_file_path = dMeasure_obj$configuration_file_path)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}

\item{configuration_file_path}{(location of SQL configuration)}
}
\value{
nothing, modifies \code{dMeasure_obj}
both these parameters have defaults, which may have
been set by previous calls
}
\description{
Opens SQL connection to SQLite configuration file.
}
\details{
Also check the SQL database
is compliant. new tables are added, and old ones
are checked to see if all required fields/columns
are present. if a field/column is missing, then
the missing field/column is added.
}
