% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{open_configuration_db}
\alias{open_configuration_db}
\title{Open the SQL connection to the configuration from the SQL configuration file}
\usage{
open_configuration_db(dMeasure_object,
  configuration_file_path = dMeasure_object$configuration_file_path,
  config_db = dMeasure_object$config_db)
}
\arguments{
\item{dMeasure_object}{dMeasure object}

\item{configuration_file_path}{(location of SQL configuration)}

\item{config_db}{the configuration DB object (R6 dbConnection object)}
}
\value{
nothing, modifies \code{dMeasure_object}
both these parameters have defaults, which may have
been set by previous calls
}
\description{
Opens SQL connection to SQLite configuration file.
}
\details{
Also check the SQL database
is compliant. new tables are added, and old ones
are checked to see if all required fields/columns
are present. if a field/column is missing, then
the missing field/column is added.
}
