% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{open_emr_db}
\alias{open_emr_db}
\title{opens the EMR database}
\usage{
open_emr_db(dMeasure_obj,
  BPdatabaseChoice = dMeasure_obj$BPdatabaseChoice,
  emr_db = dMeasure_obj$emr_db)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}

\item{BPdatabaseChoice}{the chosen database from the config_db list}

\item{emr_db}{R6 database object. this might not be initially defined

if no arguments passed, the defaults are what is stored in
the  object}
}
\description{
opens the EMR database
}
