% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{read_configuration_db}
\alias{read_configuration_db}
\title{read the SQL configuration database}
\usage{
read_configuration_db(dMeasure_obj, config_db)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}

\item{config_db}{R6 object to open SQL database
default is the internally stored value in private$config_db}
}
\description{
read the SQL configuration database
}
\examples{
dMeasure_obj <- dMeasure$new()
dMeasure_obj$open_configuration_db()
dMeasure_obj$read_configuration_db()
dMeasure_obj$UserConfig
}
