% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{read_subscription_db}
\alias{read_subscription_db}
\title{read the subscription database}
\usage{
read_subscription_db(dMeasure_obj, forcecheck = FALSE, users = NULL)
}
\arguments{
\item{dMeasure_obj}{dMeasure object}

\item{forcecheck}{check, even if already checked 'today'. TRUE/FALSE}

\item{users}{vector of user names. if NULL (the default) then all user in $UserFullConfig}
}
\description{
also update the configuration database with new Licenses (if available)
 and the date of checks
}
\examples{
dMeasure_obj$read_subscription_db()
}
