% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_server.R
\name{server.insert}
\alias{server.insert}
\title{server.insert}
\usage{
server.insert(dMeasure_obj, description)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{description}{list object : $Name, $Address, $Database, $UserID, $dbPassword

 'Address' will typically be of the form COMPUTERNAME and BPSINSTANCE, with two
 backslashes in between.

 'Address' will require 'two' backslashes in the address, because the backslash
  needs to be 'escaped' with a preceding backslash (see examples).

 'Database' will typically be BPSPATIENTS. the samples database will be BPSSAMPLES

 'UserID' will always be 'bpsrawdata'

 dbPassword will be immediately encrypted}
}
\value{
dataframe - full list of database descriptions
 can also return error (stop) if description is invalid
}
\description{
insert a new server description
}
\examples{
a <- dMeasure::dMeasure$new()
a$open_configuration_db()
a$read_configuration_db()
a$server.insert(list(Name = "MyServer", Address = "127.0.0.1\\\\BPSINSTANCE",
                     Database = "BPSSAMPLES", UserID = "bpsrawdata",
                     dbPassword = "mypassword"))
}
