% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_server.R
\name{server.update}
\alias{server.update}
\title{server.update}
\usage{
server.update(dMeasure_obj, description)
}
\arguments{
\item{dMeasure_obj}{dMeasure R6 object}

\item{description}{list $id, $Name, $Address, $Database, $UserID, $dbPassword

 any of $Name, $Address, $Database, $UserID and/or $dbPassword can
 be defined}
}
\value{
dataframe - full list of database descriptions
 can also return error (stop) if description is invalid
}
\description{
change (update) a server description
the server to change is specified by $id
}
