% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMeasure.R
\name{UserFullConfig}
\alias{UserFullConfig}
\title{UserFullConfig}
\description{
Integrates the UserConfig in the SQLite configuration file
with the user information in the EMR database
(if the clinical database is not open, then only SQLite
information is returned)
}
\details{
Contains user names attached to configuration information.
Contains ALL user names. Does NOT contain passwords.

By contrast, private$.UserConfig() contains just names
who have been configured (in SQLite), including passwords
}
